refactor socials component
